moved scaler lookup closer to its usage. Now it is hidden from other parts of the app
